+++ gnomad/assessment/sanity_checks.py
@@ -117,6 +117,9 @@ def sample_sum_check(
     :param sort_order: List containing order to sort label group combinations. Default is SORT_ORDER.
     :return: None
     """
+    if prefix != "":
+        prefix = f"{prefix}_"
+
     label_combos = make_label_combos(label_groups)
     combo_AC = [ht.info[f"{prefix}AC_{x}"] for x in label_combos]
     combo_AN = [ht.info[f"{prefix}AN_{x}"] for x in label_combos]
@@ -154,12 +157,12 @@ def sample_sum_check(
             generic_field_check(
                 ht,
                 (
-                    ht.info[f"{prefix}{subfield}_{group}_{subpop}"]
+                    ht.info[f"{prefix}{subfield}_{subpop}_{group}"]
                     != ht[f"sum_{subfield}_{group}_{alt_groups}"]
                 ),
-                f"{prefix}{subfield}_{group}_{subpop} = sum({subfield}_{group}_{alt_groups})",
+                f"{prefix}{subfield}_{subpop}_{group} = sum({subfield}_{group}_{alt_groups})",
                 [
-                    f"info.{prefix}{subfield}_{group}_{subpop}",
+                    f"info.{prefix}{subfield}_{subpop}_{group}",
                     f"sum_{subfield}_{group}_{alt_groups}",
                 ],
                 verbose,